requirements.txt
Django==4.2.5
django_multiupload==0.6.1
firebase_admin==6.2.0
Pyrebase4==4.7.1
weasyprint==60.1
